# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto APIs
  description: |-
    This is the documentation for the APIs for the WASAPhoto Project,
    for the accademic year 23/24 by Scutigliani Luca (1914401)
  version: 1.1.0
servers:
  - url: https://wasaphoto.lucascutigliani.it
    description: Production URL
  - url: https://wasadev.lucascutigliani.it
    description: Development URL
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT"
  schemas:
...
    UUID:
      description: |
        Universally unique identifier of the user.
      type: string
      pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
      minLength: 36
      maxLength: 36

    # Counters
...
    Photo:
      description: |
        This component is the representation
        of a photo.

        It contains a simple id for the photo,
        the UUID of the author, the date of
        upload, the likes/comments counters
        and whether or not the photo has been
        liked from the pov of the user executing
        an operation.
      type: object
      properties:
        id:
          { $ref: "#/components/schemas/UUID" }
        author:
          { $ref: "#/components/schemas/UUID" }
        date:
          { $ref: "#/components/schemas/Date" }
        extension:
          description: Photo extension (png, jpg)
          type: string
          minLength: 3
          maxLength: 10
          enum: [png, jpg]
        likesCount:
          description: Counter for photo likes.
          type: integer
          minimum: 0
          example: 42
        commentsCount:
          description: Counter for photo comments.
          type: integer
          minimum: 0
          example: 42
        liked:
          description: |
            Flag indicating if the requesting user has
            liked the photo.
          type: boolean
          example: false
      readOnly: true
    PhotoURL:
      description: Components representing the url of an uploaded photo.
      readOnly: true
      type: string
      pattern: '^/uploads/[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}/[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}\.(png|jpg)$'
      example: "/uploads/e5f04445-2f4a-4e14-a8fb-5d7f84221096/9f70128d-8cf0-3892-9e51-3980525107a5.png"
...
    Date:
      description: Simple date representation.
      type: string
      format: date-time
      readOnly: true
      example: 2024-01-16 17:49:43.640624
  parameters:
...
    paginationIndex:
      name: paginationIndex
      in: query
      required: false
      description: Current page index.
      schema:
        type: integer
        minimum: 0
        example: 2
    photoID:
      name: photoID
      in: path
      required: true
      description: Photo identifier parameter.
      schema:
        { $ref: "#/components/schemas/UUID" }
...
    X-Requesting-User-UUID:
      name: X-Requesting-User-UUID
      in: header
      required: true
      description: UUID of the requesting user.
      schema:
        { $ref: "#/components/schemas/UUID" }
  responses:
    BadRequest:
      description: |
        Bad Request. The request syntax is incorrect.
      content:
        text/plain:
          schema:
            description: Specific error message for error 400.
            type: string
            example: |
              Bad Request: Syntax error in the request.
    Unauthorized:
      description: |
        Unauthorized. Authentication is required and
        has failed or has not been provided.
      content:
        text/plain:
          schema:
            description: Specific error message for error 401.
            type: string
            example: |
              Unauthorized: Authentication failed or not provided.
    Forbidden:
      description: |
        Forbidden. The user does not have the
        permissions to perform this action.
      content:
        text/plain:
          schema:
            description: Specific error message for error 403.
            type: string
            example: |
              Forbidden: Insufficient permissions to perform
              this action. You may have tried to perform an
              operation on another user data or tried to
              retrieve data from a user that banned you.
    NotFound:
      description: |
        Not Found. The requested resource could
        not be found.
      content:
        text/plain:
          schema:
            description: Specific error message for error 404.
            type: string
            example: |
              Not Found: The requested resource was not found.
    InternalServerError:
      description: |
        Internal Server Error. A generic error
        occurred on the server.
      content:
        text/plain:
          schema:
            description: Specific error message for error 500.
            type: string
            example: |
              Internal Server Error: A generic error
              occurred on the server.
paths:
...
  /photos:
    parameters:
      - { $ref: "#/components/parameters/X-Requesting-User-UUID" }
    get:
      parameters:
        - name: userID
          in: query
          description: User UUID
          required: true
          schema:
            { $ref: "#/components/schemas/UUID" }
        - { $ref: "#/components/parameters/paginationIndex" }
      tags: ["Photos"]
      operationId: getPhotos
      security:
        - BearerAuth: []
      summary: Get user photos
      description: |
        This endpoint is used to get the photos
        of a specific user by it's UUID.

        Will be returned 10 photos at a time,
        ordered from the most recent.
        To get more images, you must increment the
        "paginationIndex" parameter by 1 and so on until
        the given "paginationLimit".

        If the requested userID does not exist, or the "paginationIndex"
        is bigger than "paginationLimit" an error will be returned.
      responses:
        '200':
          description: User photos in the content.
          content:
            application/json:
              schema:
                description: Object containing the user photos ids
                type: object
                properties:
                  user-photos:
                    description: |
                      Array of photo ids
                    type: array
                    minItems: 0
                    maxItems: 10
                    items:
                      { $ref: "#/components/schemas/UUID" }
                  paginationLimit:
                    description: |
                      The paginationLimit represents the maximum
                      value for paginationIndex
                    type: integer
                    minimum: 0
                    example: 4
        '400':
          { $ref: "#/components/responses/BadRequest" }
        '401':
          { $ref: "#/components/responses/Unauthorized" }
        '403':
          { $ref: "#/components/responses/Forbidden" }
        '404':
          { $ref: "#/components/responses/NotFound" }
        '500':
          { $ref: "#/components/responses/InternalServerError" }
    post:
      tags: ["Photos"]
      operationId: uploadPhoto
      security:
        - BearerAuth: []
      summary: Upload a photo
      description: |
        This endpoint is used to upload a photo.
        If an error occurs, a error is returned.
      x-codegen-request-body-name: photoUpload
      requestBody:
        required: true
        description: |
          The request body should contain
          the image that needs to be uploaded
          in binary format and the UUID of the
          user uploading the photo.
        content:
          image/*:
            schema:
              description: |
                To calculate max/min length for the image
                we can approximate the maximum teorical
                dimension for the most lossless format
                we will accept (in this case PNG).
                In this case, we can assume that we want
                only images that are smaller than 4032x4032
                pixels, thus leading us to this calculation

                4032 pixel x 4032 pixel x 4 byte/pixel = around 65 MB

                With this in mind, we can safely say that all
                the images the we will receive will be smaller
                than 65.000.000 bytes.

                We will set a 100 bytes minimum length to ensure
                that data is present.
              type: string
              format: binary
              minLength: 100
              maxLength: 65000000
      responses:
        '201':
          description: Photo successfully uploaded, photo id in the content
          headers:
            Location:
              description: |
                URL location of the uploaded image.
              schema:
                { $ref: "#/components/schemas/PhotoURL" }
          content:
            application/json:
              schema:
                { $ref: "#/components/schemas/Photo" }
        '400':
          { $ref: "#/components/responses/BadRequest" }
        '401':
          { $ref: "#/components/responses/Unauthorized" }
        '403':
          { $ref: "#/components/responses/Forbidden" }
        '404':
          { $ref: "#/components/responses/NotFound" }
        '500':
          { $ref: "#/components/responses/InternalServerError" }
  /photos/{photoID}:
    parameters:
      - { $ref: "#/components/parameters/photoID" }
      - { $ref: "#/components/parameters/X-Requesting-User-UUID" }
    get:
      tags: ["Photos"]
      operationId: getPhoto
      security:
        - BearerAuth: []
      summary: Get photo info
      description: |
        This endpoint is used to get
        infos about a photo.
      responses:
        '200':
          description: Photo info in the content.
          content:
            application/json:
              schema:
                { $ref: "#/components/schemas/Photo" }
        '400':
          { $ref: "#/components/responses/BadRequest" }
        '401':
          { $ref: "#/components/responses/Unauthorized" }
        '403':
          { $ref: "#/components/responses/Forbidden" }
        '404':
          { $ref: "#/components/responses/NotFound" }
        '500':
          { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["Photos"]
      operationId: deletePhoto
      security:
        - BearerAuth: []
      summary: Delete a photo
      description: |
        This endpoint allows the user to delete
        one of its photos using the ID.
      responses:
        '204':
          description: The photo has been deleted successfully.
        '400':
          { $ref: "#/components/responses/BadRequest" }
        '401':
          { $ref: "#/components/responses/Unauthorized" }
        '403':
          { $ref: "#/components/responses/Forbidden" }
        '404':
          { $ref: "#/components/responses/NotFound" }
        '500':
          { $ref: "#/components/responses/InternalServerError" }
...